fix: resolve Linux-only clippy errors for CI
- Add Default impl for BreakpointManager and WatchpointManager
  (next_id starts at 1, so manual impl instead of derive)
- Remove redundant closure in format_stack_frame mapping
- Remove unnecessary i64/i32 casts in dap_server and process
- Simplify Ok(result?) to result in target.rs

+++ src/process.rs
@@ -317,7 +317,7 @@ impl Process {
                 }
                 WaitStatus::PtraceEvent(tid, _, event) => {
                     self.state = ProcessState::Stopped;
-                    if event == libc::PTRACE_EVENT_CLONE as i32 {
+                    if event == libc::PTRACE_EVENT_CLONE {
                         let new_pid_raw = ptrace::getevent(tid)
                             .map_err(|e| Error::Process(format!("getevent: {}", e)))?;
                         let new_tid = Pid::from_raw(new_pid_raw as i32);
